Create a new component for transaction button
