+++ ui/components/Signing/Signer/SignerBaseFrame.tsx
@@ -1,11 +1,8 @@
-import React, { ReactElement, useEffect, useRef, useState } from "react"
+import React, { ReactElement } from "react"
 import { useTranslation } from "react-i18next"
-import {
-  selectIsTransactionLoaded,
-  selectTransactionData,
-} from "@tallyho/tally-background/redux-slices/selectors/transactionConstructionSelectors"
-import { useBackgroundSelector, useDebounce } from "../../../hooks"
-import SharedButton from "../../Shared/SharedButton"
+import { selectTransactionData } from "@tallyho/tally-background/redux-slices/selectors/transactionConstructionSelectors"
+import { useBackgroundSelector } from "../../../hooks"
+import TransactionButton from "./TransactionButton"
 
 type SignerBaseFrameProps = {
   signingActionLabel: string
@@ -22,88 +19,27 @@ export default function SignerBaseFrame({
 }: SignerBaseFrameProps): ReactElement {
   const { t } = useTranslation("translation", { keyPrefix: "signTransaction" })
   const transactionDetails = useBackgroundSelector(selectTransactionData)
-
-  const isTransactionDataReady = useBackgroundSelector(
-    selectIsTransactionLoaded
-  )
   const hasInsufficientFunds =
     transactionDetails?.annotation?.warnings?.includes("insufficient-funds")
-
-  /*
-    Prevent shenanigans by disabling the sign button for a bit
-    when rendering new sign content or when changing window focus.
-  */
-  const delaySignButtonTimeout = useRef<number | undefined>()
-
-  const [isOnDelayToSign, setIsOnDelayToSign] = useState(false)
-  // Debounced unlock buttons because dispatching transaction events is async and can happen in batches
-  const [unlockButtons, setUnlockButtons] = useDebounce(
-    isTransactionDataReady,
-    300
-  )
-  const [focusChangeNonce, setFocusChangeNonce] = useState(0)
-
-  function clearDelaySignButtonTimeout() {
-    if (typeof delaySignButtonTimeout.current !== "undefined") {
-      clearTimeout(delaySignButtonTimeout.current)
-      delaySignButtonTimeout.current = undefined
-    }
-  }
-
-  useEffect(
-    () => setUnlockButtons(isTransactionDataReady),
-    [isTransactionDataReady, setUnlockButtons]
-  )
-
-  useEffect(() => {
-    const increaseFocusChangeNonce = () => {
-      setFocusChangeNonce((x) => x + 1)
-    }
-    window.addEventListener("focus", increaseFocusChangeNonce)
-    window.addEventListener("blur", increaseFocusChangeNonce)
-
-    return () => {
-      window.removeEventListener("focus", increaseFocusChangeNonce)
-      window.removeEventListener("blur", increaseFocusChangeNonce)
-    }
-  }, [])
-
-  // Runs on updates
-  useEffect(() => {
-    clearDelaySignButtonTimeout()
-
-    if (document.hasFocus()) {
-      delaySignButtonTimeout.current = window.setTimeout(() => {
-        setIsOnDelayToSign(false)
-        // Random delay between 0.5 and 2 seconds
-      }, Math.floor(Math.random() * (5 - 1) + 1) * 500)
-    } else {
-      setIsOnDelayToSign(true)
-    }
-  }, [focusChangeNonce])
 
   return (
     <>
       <div className="signature-details">{children}</div>
       <footer>
-        <SharedButton
-          size="large"
-          type="secondary"
-          onClick={onReject}
-          isDisabled={!unlockButtons}
-        >
+        <TransactionButton size="large" type="secondary" onClick={onReject}>
           {t("reject")}
-        </SharedButton>
+        </TransactionButton>
 
-        <SharedButton
+        <TransactionButton
           type="primaryGreen"
           size="large"
           onClick={onConfirm}
+          isDisabled={hasInsufficientFunds}
           showLoadingOnClick
-          isDisabled={isOnDelayToSign || !unlockButtons || hasInsufficientFunds}
+          reactWindowFocus
         >
           {signingActionLabel}
-        </SharedButton>
+        </TransactionButton>
       </footer>
       <style jsx>
         {`
